fix: config file not found (#390)

+++ images/utils/launcher/config/config.py
@@ -402,9 +402,8 @@ class ConfigLoader:
     def load_general_config(self, home_dir):
         config_file = get_hostfs_file(f"{home_dir}/xud-docker.conf")
         sample_config_file = get_hostfs_file(f"{home_dir}/sample-xud-docker.conf")
-        try:
-            copyfile(os.path.dirname(__file__) + "/xud-docker.conf", sample_config_file)
-        except FileNotFoundError:
+        copyfile(os.path.dirname(__file__) + "/xud-docker.conf", sample_config_file)
+        if not os.path.exists(config_file):
             copyfile(os.path.dirname(__file__) + f"/xud-docker.conf", config_file)
         with open(config_file) as f:
             return f.read()
@@ -412,9 +411,8 @@ class ConfigLoader:
     def load_network_config(self, network, network_dir):
         config_file = get_hostfs_file(f"{network_dir}/{network}.conf")
         sample_config_file = get_hostfs_file(f"{network_dir}/sample-{network}.conf")
-        try:
-            copyfile(os.path.dirname(__file__) + f'/{network}.conf', sample_config_file)
-        except FileNotFoundError:
+        copyfile(os.path.dirname(__file__) + f'/{network}.conf', sample_config_file)
+        if not os.path.exists(config_file):
             copyfile(os.path.dirname(__file__) + f"/{network}.conf", config_file)
         with open(config_file) as f:
             return f.read()
